+++ src/main/java/com/clubber/ClubberServer/domain/recruit/dto/recruitCalendar/CreateLinkedCalendarRequest.java
@@ -1,4 +1,4 @@
-package com.clubber.ClubberServer.domain.calendar.dto;
+package com.clubber.ClubberServer.domain.recruit.dto.recruitCalendar;
 
 public record CreateLinkedCalendarRequest(
         Long recruitId,
